# doc/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "S4U Rest API",
        "description": "API REST for user and ticket management on S4U App",
        "termsOfService": "",
        "contact": {
            "name": "Julian Avellaneda",
            "url": "http://www.uwasoftco.com",
            "email": "julian.avellaneda@uwasoftco.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "servers": [{
            "url": "http://localhost:3001/v1/s4u-api",
            "description": "Local Dev Server"
        },
        {
            "url": "https://s4uapi.herokuapp.com/v1/s4u-api",
            "description": "Cloud Dev Server"
        }
    ],
    "tags": [{
        "name": "Ticket",
        "description": "Everything about tickets management"
    }],
    "paths": {
        "/ticket": {
            "get": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Get tickets",
                "operationId": "getTickets",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ticket"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Update an existing ticket",
                "operationId": "updateTicket",
                "requestBody": {
                    "description": "Ticket object that needs to be added to the systems",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Ticket"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Ticket updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "x-codegen-request-body-name": "body"
            },
            "post": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Add a new ticket to the systems",
                "operationId": "createTicket",
                "requestBody": {
                    "description": "Ticket object that needs to be added to the systems",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Ticket"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Ticket created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "x-codegen-request-body-name": "body"
            }
        },
        "/{id}": {
            "get": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Find ticket by ID",
                "description": "Returns a single ticket",
                "operationId": "getTicketById",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ID of ticket to return",
                    "required": true,
                    "schema": {
                        "type": "integer",
                        "format": "int64"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user-owner/{id}": {
            "get": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Find tickets by user owner ID",
                "description": "Returns all tickets for the specified user owner",
                "operationId": "getTicketsByOwnerUser",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ID of the user",
                    "required": true,
                    "schema": {
                        "type": "integer",
                        "format": "int64"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ticket"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pet not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user-assigned/{id}": {
            "get": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Find tickets by assigned user ID",
                "description": "Returns all tickets for the specified assigned user",
                "operationId": "getTicketsByAssignedUser",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ID of the user",
                    "required": true,
                    "schema": {
                        "type": "integer",
                        "format": "int64"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ticket"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pet not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Ticket": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "priority": {
                        "type": "object",
                        "properties": {
                            "text": {
                                "type": "string"
                            },
                            "code": {
                                "type": "number"
                            }
                        }
                    },
                    "status": {
                        "type": "object",
                        "properties": {
                            "text": {
                                "type": "string"
                            },
                            "code": {
                                "type": "number"
                            }
                        }
                    },
                    "type": {
                        "type": "object",
                        "properties": {
                            "text": {
                                "type": "string"
                            },
                            "code": {
                                "type": "number"
                            }
                        }
                    },
                    "category": {
                        "type": "object",
                        "properties": {
                            "text": {
                                "type": "string"
                            },
                            "code": {
                                "type": "number"
                            }
                        }
                    },
                    "userOwner": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "userAssinged": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "ApiResponse": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "type": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            }
        }
    }
}